requirements.txt
https://github.com/msavva/socketIO-client-2/zipball/master
numpy
scipy
easydict
pygame
Pillow
psutil
pyyaml
